Move common parts of Channel out of Async module
