+++ src/Streamly/Internal/Data/Stream/Async/Channel/Dispatcher.hs
@@ -9,22 +9,8 @@
 --
 module Streamly.Internal.Data.Stream.Async.Channel.Dispatcher
     (
-    -- * Latency collection
-      collectLatency
-
-    -- * Diagnostics
-    , withDiagMVar
-    , dumpSVar
-    , printSVar
-
-    -- * Thread accounting
-    , delThread
-    , modifyThread
-    , allThreadsDone
-
     -- * Dispatching
-    , recordMaxWorkers
-    , pushWorker
+      pushWorker
     , dispatchWorker
     , dispatchWorkerPaced
     , sendWorkerWait
@@ -35,338 +21,33 @@ module Streamly.Internal.Data.Stream.Async.Channel.Dispatcher
 where
 
 import Control.Monad.Trans.Control (MonadBaseControl)
-import Control.Concurrent (takeMVar, tryReadMVar, ThreadId, threadDelay)
-import Control.Concurrent.MVar (tryPutMVar)
-import Control.Exception
-       (assert, catches, throwIO, Handler(..), BlockedIndefinitelyOnMVar(..),
-        BlockedIndefinitelyOnSTM(..))
+import Control.Concurrent (takeMVar, threadDelay)
+import Control.Exception (assert)
 import Control.Monad (when, void)
 import Control.Monad.IO.Class (MonadIO(liftIO))
 import Data.Maybe (fromJust, fromMaybe)
-import Data.IORef (IORef, modifyIORef, newIORef, readIORef, writeIORef)
+import Data.IORef (modifyIORef, newIORef, readIORef, writeIORef)
 import Streamly.Internal.Control.ForkLifted (doFork)
-import Streamly.Internal.Data.Atomics
-       (atomicModifyIORefCAS, atomicModifyIORefCAS_, writeBarrier,
-        storeLoadBarrier)
+import Streamly.Internal.Data.Atomics (atomicModifyIORefCAS_, storeLoadBarrier)
 import Streamly.Internal.Data.Time.Clock (Clock(Monotonic), getTime)
 import Streamly.Internal.Data.Time.Units
-       (AbsTime, NanoSecond64(..), MicroSecond64(..), diffAbsTime64,
-        fromRelTime64, toRelTime64, showNanoSecond64, showRelTime64)
-import System.IO (hPutStrLn, stderr)
-
--- import qualified Data.Heap as H
-import qualified Data.Set as S
+       (MicroSecond64(..), diffAbsTime64, fromRelTime64, toRelTime64)
 
 import Streamly.Internal.Data.Stream.Async.Channel.Type
-import Streamly.Internal.Data.Stream.Async.Channel.Worker
-
--------------------------------------------------------------------------------
--- Worker latency data processing
--------------------------------------------------------------------------------
-
--- Every once in a while workers update the latencies and check the yield rate.
--- They return if we are above the expected yield rate. If we check too often
--- it may impact performance, if we check less often we may have a stale
--- picture. We update every minThreadDelay but we translate that into a yield
--- count based on latency so that the checking overhead is little.
---
--- XXX use a generation count to indicate that the value is updated. If the
--- value is updated an existing worker must check it again on the next yield.
--- Otherwise it is possible that we may keep updating it and because of the mod
--- worker keeps skipping it.
-updateWorkerPollingInterval :: YieldRateInfo -> NanoSecond64 -> IO ()
-updateWorkerPollingInterval yinfo latency = do
-    let periodRef = workerPollingInterval yinfo
-        cnt = max 1 $ minThreadDelay `div` latency
-        period = min cnt (fromIntegral magicMaxBuffer)
-
-    writeIORef periodRef (fromIntegral period)
-
-{-# INLINE recordMinMaxLatency #-}
-recordMinMaxLatency :: Channel m a -> NanoSecond64 -> IO ()
-recordMinMaxLatency sv new = do
-    let ss = svarStats sv
-    minLat <- readIORef (minWorkerLatency ss)
-    when (new < minLat || minLat == 0) $
-        writeIORef (minWorkerLatency ss) new
-
-    maxLat <- readIORef (maxWorkerLatency ss)
-    when (new > maxLat) $ writeIORef (maxWorkerLatency ss) new
-
-recordAvgLatency :: Channel m a -> (Count, NanoSecond64) -> IO ()
-recordAvgLatency sv (count, time) = do
-    let ss = svarStats sv
-    modifyIORef (avgWorkerLatency ss) $
-        \(cnt, t) -> (cnt + count, t + time)
-
--- Pour the pending latency stats into a collection bucket
-{-# INLINE collectWorkerPendingLatency #-}
-collectWorkerPendingLatency
-    :: IORef (Count, Count, NanoSecond64)
-    -> IORef (Count, Count, NanoSecond64)
-    -> IO (Count, Maybe (Count, NanoSecond64))
-collectWorkerPendingLatency cur col = do
-    (fcount, count, time) <- atomicModifyIORefCAS cur $ \v -> ((0,0,0), v)
-
-    (fcnt, cnt, t) <- readIORef col
-    let totalCount = fcnt + fcount
-        latCount   = cnt + count
-        latTime    = t + time
-    writeIORef col (totalCount, latCount, latTime)
-
-    assert (latCount == 0 || latTime /= 0) (return ())
-    let latPair =
-            if latCount > 0 && latTime > 0
-            then Just (latCount, latTime)
-            else Nothing
-    return (totalCount, latPair)
-
-{-# INLINE shouldUseCollectedBatch #-}
-shouldUseCollectedBatch
-    :: Count
-    -> NanoSecond64
-    -> NanoSecond64
-    -> NanoSecond64
-    -> Bool
-shouldUseCollectedBatch collectedYields collectedTime newLat prevLat =
-    let r = fromIntegral newLat / fromIntegral prevLat :: Double
-    in     (collectedYields > fromIntegral magicMaxBuffer)
-        || (collectedTime > minThreadDelay)
-        || (prevLat > 0 && (r > 2 || r < 0.5))
-        || (prevLat == 0)
-
--- Returns a triple, (1) yield count since last collection, (2) the base time
--- when we started counting, (3) average latency in the last measurement
--- period. The former two are used for accurate measurement of the going rate
--- whereas the average is used for future estimates e.g. how many workers
--- should be maintained to maintain the rate.
--- CAUTION! keep it in sync with getWorkerLatency
-collectLatency :: Channel m a
-               -> YieldRateInfo
-               -> Bool
-               -> IO (Count, AbsTime, NanoSecond64)
-collectLatency sv yinfo drain = do
-    let cur      = workerPendingLatency yinfo
-        col      = workerCollectedLatency yinfo
-        longTerm = svarAllTimeLatency yinfo
-        measured = workerMeasuredLatency yinfo
-
-    (newCount, newLatPair) <- collectWorkerPendingLatency cur col
-    (lcount, ltime) <- readIORef longTerm
-    prevLat <- readIORef measured
-
-    let newLcount = lcount + newCount
-        retWith lat = return (newLcount, ltime, lat)
-
-    case newLatPair of
-        Nothing -> retWith prevLat
-        Just (count, time) -> do
-            let newLat = time `div` fromIntegral count
-            when (svarInspectMode sv) $ recordMinMaxLatency sv newLat
-            -- When we have collected a significant sized batch we compute the
-            -- new latency using that batch and return the new latency,
-            -- otherwise we return the previous latency derived from the
-            -- previous batch.
-            if shouldUseCollectedBatch newCount time newLat prevLat || drain
-            then do
-                -- XXX make this NOINLINE?
-                updateWorkerPollingInterval yinfo (max newLat prevLat)
-                when (svarInspectMode sv) $ recordAvgLatency sv (count, time)
-                writeIORef col (0, 0, 0)
-                writeIORef measured ((prevLat + newLat) `div` 2)
-                modifyIORef longTerm $ \(_, t) -> (newLcount, t)
-                retWith newLat
-            else retWith prevLat
-
--------------------------------------------------------------------------------
--- Dumping the SVar for debug/diag
--------------------------------------------------------------------------------
-
-dumpSVarStats :: Channel m a -> SVarStats -> IO String
-dumpSVarStats sv ss = do
-    case yieldRateInfo sv of
-        Nothing -> return ()
-        Just yinfo -> do
-            _ <- liftIO $ collectLatency sv yinfo True
-            return ()
-
-    dispatches <- readIORef $ totalDispatches ss
-    maxWrk <- readIORef $ maxWorkers ss
-    maxOq <- readIORef $ maxOutQSize ss
-    -- maxHp <- readIORef $ maxHeapSize ss
-    minLat <- readIORef $ minWorkerLatency ss
-    maxLat <- readIORef $ maxWorkerLatency ss
-    (avgCnt, avgTime) <- readIORef $ avgWorkerLatency ss
-    (svarCnt, svarGainLossCnt, svarLat) <- case yieldRateInfo sv of
-        Nothing -> return (0, 0, 0)
-        Just yinfo -> do
-            (cnt, startTime) <- readIORef $ svarAllTimeLatency yinfo
-            if cnt > 0
-            then do
-                t <- readIORef (svarStopTime ss)
-                gl <- readIORef (svarGainedLostYields yinfo)
-                case t of
-                    Nothing -> do
-                        now <- getTime Monotonic
-                        let interval = diffAbsTime64 now startTime
-                        return (cnt, gl, interval `div` fromIntegral cnt)
-                    Just stopTime -> do
-                        let interval = diffAbsTime64 stopTime startTime
-                        return (cnt, gl, interval `div` fromIntegral cnt)
-            else return (0, 0, 0)
-
-    return $ unlines
-        [ "total dispatches = " <> show dispatches
-        , "max workers = " <> show maxWrk
-        , "max outQSize = " <> show maxOq
-            <> (if minLat > 0
-               then "\nmin worker latency = " <> showNanoSecond64 minLat
-               else "")
-            <> (if maxLat > 0
-               then "\nmax worker latency = " <> showNanoSecond64 maxLat
-               else "")
-            <> (if avgCnt > 0
-                then let lat = avgTime `div` fromIntegral avgCnt
-                     in "\navg worker latency = " <> showNanoSecond64 lat
-                else "")
-            <> (if svarLat > 0
-               then "\nSVar latency = " <> showRelTime64 svarLat
-               else "")
-            <> (if svarCnt > 0
-               then "\nSVar yield count = " <> show svarCnt
-               else "")
-            <> (if svarGainLossCnt > 0
-               then "\nSVar gain/loss yield count = " <> show svarGainLossCnt
-               else "")
-        ]
-
-{-# NOINLINE dumpSVar #-}
-dumpSVar :: Channel m a -> IO String
-dumpSVar sv = do
-    (oqList, oqLen) <- readIORef $ outputQueue sv
-    db <- tryReadMVar $ outputDoorBell sv
-
-    waiting <- readIORef $ needDoorBell sv
-    rthread <- readIORef $ workerThreads sv
-    workers <- readIORef $ workerCount sv
-    stats <- dumpSVarStats sv (svarStats sv)
-
-    return $ unlines
-        [
-          "Creator tid = " <> show (svarCreator sv)
-        , "---------CURRENT STATE-----------"
-        , "outputQueue length computed  = " <> show (length oqList)
-        , "outputQueue length maintained = " <> show oqLen
-        -- XXX print the types of events in the outputQueue, first 5
-        , "outputDoorBell = " <> show db
-        ]
-        <> unlines
-        [ "needDoorBell = " <> show waiting
-        , "running threads = " <> show rthread
-        -- XXX print the status of first 5 threads
-        , "running thread count = " <> show workers
-        ]
-        <> "---------STATS-----------\n"
-        <> stats
-
--- MVar diagnostics has some overhead - around 5% on AsyncT null benchmark, we
--- can keep it on in production to debug problems quickly if and when they
--- happen, but it may result in unexpected output when threads are left hanging
--- until they are GCed because the consumer went away.
-
-{-# NOINLINE mvarExcHandler #-}
-mvarExcHandler :: Channel m a -> String -> BlockedIndefinitelyOnMVar -> IO ()
-mvarExcHandler sv label e@BlockedIndefinitelyOnMVar = do
-    svInfo <- dumpSVar sv
-    hPutStrLn stderr $ label <> " " <> "BlockedIndefinitelyOnMVar\n" <> svInfo
-    throwIO e
-
-{-# NOINLINE stmExcHandler #-}
-stmExcHandler :: Channel m a -> String -> BlockedIndefinitelyOnSTM -> IO ()
-stmExcHandler sv label e@BlockedIndefinitelyOnSTM = do
-    svInfo <- dumpSVar sv
-    hPutStrLn stderr $ label <> " " <> "BlockedIndefinitelyOnSTM\n" <> svInfo
-    throwIO e
-
-withDiagMVar :: Channel m a -> String -> IO () -> IO ()
-withDiagMVar sv label action =
-    if svarInspectMode sv
-    then
-        action `catches` [ Handler (mvarExcHandler sv label)
-                         , Handler (stmExcHandler sv label)
-                         ]
-    else action
-
-printSVar :: Channel m a -> String -> IO ()
-printSVar sv how = do
-    svInfo <- dumpSVar sv
-    hPutStrLn stderr $ "\n" <> how <> "\n" <> svInfo
-
--------------------------------------------------------------------------------
--- Thread accounting
--------------------------------------------------------------------------------
-
--- Thread tracking is needed for two reasons:
---
--- 1) Killing threads on exceptions. Threads may not be left to go away by
--- themselves because they may run for significant times before going away or
--- worse they may be stuck in IO and never go away.
---
--- 2) To know when all threads are done and the stream has ended.
-
-{-# NOINLINE addThread #-}
-addThread :: MonadIO m => Channel m a -> ThreadId -> m ()
-addThread sv tid =
-    liftIO $ modifyIORef (workerThreads sv) (S.insert tid)
-
--- This is cheaper than modifyThread because we do not have to send a
--- outputDoorBell This can make a difference when more workers are being
--- dispatched.
-{-# INLINE delThread #-}
-delThread :: MonadIO m => Channel m a -> ThreadId -> m ()
-delThread sv tid =
-    liftIO $ modifyIORef (workerThreads sv) (S.delete tid)
-
--- If present then delete else add. This takes care of out of order add and
--- delete i.e. a delete arriving before we even added a thread.
--- This occurs when the forked thread is done even before the 'addThread' right
--- after the fork gets a chance to run.
-{-# INLINE modifyThread #-}
-modifyThread :: MonadIO m => Channel m a -> ThreadId -> m ()
-modifyThread sv tid = do
-    changed <- liftIO $ atomicModifyIORefCAS (workerThreads sv) $ \old ->
-        if S.member tid old
-        then let new = S.delete tid old in (new, new)
-        else let new = S.insert tid old in (new, old)
-    when (null changed) $
-         liftIO $ do
-            writeBarrier
-            void $ tryPutMVar (outputDoorBell sv) ()
-
--- | This is safe even if we are adding more threads concurrently because if
--- a child thread is adding another thread then anyway 'workerThreads' will
--- not be empty.
-{-# INLINE allThreadsDone #-}
-allThreadsDone :: MonadIO m => Channel m a -> m Bool
-allThreadsDone sv = liftIO $ S.null <$> readIORef (workerThreads sv)
+import Streamly.Internal.Data.Stream.Channel.Dispatcher
+import Streamly.Internal.Data.Stream.Channel.Types
+import Streamly.Internal.Data.Stream.Channel.Worker
 
 -------------------------------------------------------------------------------
 -- Dispatching workers
 -------------------------------------------------------------------------------
 
-{-# NOINLINE recordMaxWorkers #-}
-recordMaxWorkers :: MonadIO m => Channel m a -> m ()
-recordMaxWorkers sv = liftIO $ do
-    active <- readIORef (workerCount sv)
-    maxWrk <- readIORef (maxWorkers $ svarStats sv)
-    when (active > maxWrk) $ writeIORef (maxWorkers $ svarStats sv) active
-    modifyIORef (totalDispatches $ svarStats sv) (+1)
-
 {-# NOINLINE pushWorker #-}
 pushWorker :: (MonadIO m, MonadBaseControl IO m) => Count -> Channel m a -> m ()
 pushWorker yieldMax sv = do
     liftIO $ atomicModifyIORefCAS_ (workerCount sv) $ \n -> n + 1
-    when (svarInspectMode sv) $ recordMaxWorkers sv
+    when (svarInspectMode sv)
+        $ recordMaxWorkers (workerCount sv) (svarStats sv)
     -- This allocation matters when significant number of workers are being
     -- sent. We allocate it only when needed.
     winfo <-
@@ -381,8 +62,12 @@ pushWorker yieldMax sv = do
                     , workerYieldCount = cntRef
                     , workerLatencyStart = lat
                     }
-    doFork (workLoop sv winfo) (svarMrun sv) (handleChildException sv)
-        >>= addThread sv
+    doFork (workLoop sv winfo) (svarMrun sv) exception
+        >>= addThread (workerThreads sv)
+
+    where
+
+    exception = handleChildException (outputQueue sv) (outputDoorBell sv)
 
 -- Returns:
 -- True: can dispatch more
@@ -454,7 +139,9 @@ dispatchWorkerPaced sv = do
     (svarYields, svarElapsed, wLatency) <- do
         now <- liftIO $ getTime Monotonic
         (yieldCount, baseTime, lat) <-
-            liftIO $ collectLatency sv yinfo False
+            liftIO
+                $ collectLatency
+                    (svarInspectMode sv) (svarStats sv) yinfo False
         let elapsed = fromRelTime64 $ diffAbsTime64 now baseTime
         let latency =
                 if lat == 0
@@ -492,7 +179,7 @@ dispatchWorkerPaced sv = do
                 -- here as it is not a reliable way to ensure there are
                 -- definitely no active workers. When workerCount is 0 we may
                 -- still have a Stop event waiting in the outputQueue.
-                done <- allThreadsDone sv
+                done <- allThreadsDone (workerThreads sv)
                 when done $ void $ do
                     let us = fromRelTime64 (toRelTime64 s) :: MicroSecond64
                     liftIO $ threadDelay (fromIntegral us)
@@ -502,7 +189,7 @@ dispatchWorkerPaced sv = do
                 assert (yields > 0) (return ())
                 updateGainedLostYields yinfo yields
 
-                done <- allThreadsDone sv
+                done <- allThreadsDone (workerThreads sv)
                 when done $ void $ dispatchWorker yields sv
                 return False
             ManyWorkers netWorkers yields -> do
@@ -627,8 +314,12 @@ sendWorkerWait delay dispatch sv = do
         if canDoMore
         then sendWorkerWait delay dispatch sv
         else do
-            liftIO $ withDiagMVar sv "sendWorkerWait: nothing to do"
-                             $ takeMVar (outputDoorBell sv)
+            liftIO
+                $ withDiagMVar
+                    (svarInspectMode sv)
+                    (dumpSVar sv)
+                    "sendWorkerWait: nothing to do"
+                $ takeMVar (outputDoorBell sv)
             (_, len) <- liftIO $ readIORef (outputQueue sv)
             when (len <= 0) $ sendWorkerWait delay dispatch sv
